Implemented better saving and logging schemes Added loss plots Implemented a function to preprocess a single image

+++ trainer.py
@@ -2,6 +2,8 @@ import torch
 import time
 import util
 import json
+import os
+from os.path import join
 
 
 def train(optimiser, model, data_loader, device, loss_fn):
@@ -45,22 +47,41 @@ def train_and_evaluate(epoch, device, training_dict, n_class):
     training_dict['lr_scheduler'].step()
     elapsed_time = time.time() - start
 
-    util.generate_example_output(training_dict['model'], training_dict['dataset'], device, epoch, n_class,
-                                 dims=training_dict['dims'])
+    # only save every 10 epochs
+    if (epoch + 1) % 10 == 0:
+        util.generate_example_output(training_dict, device, epoch, n_class)
+        torch.save({
+            'epoch': epoch,
+            'model_state_dict': training_dict['model'].state_dict(),
+            'optimiser_state_dict': training_dict['optimiser'].state_dict(),
+            'scheduler_state_dict': training_dict['lr_scheduler'].state_dict(),
+            'train_loss': training_dict['train_logger'],
+            'valid_loss': training_dict['validation_logger']
+        }, training_dict['model_path'])
 
-    with open(f'out/loss{training_dict["dims"]}D.json', 'w') as outfile:
+    with open(join(training_dict['out_path'], f'loss{training_dict["dims"]}D.json'), 'w') as outfile:
         json.dump({'epochs_completed': epoch,
                    'train': training_dict['train_logger'],
                    'validation': training_dict['validation_logger']}, outfile, indent=4)
 
-    torch.save({
-        'epoch': epoch,
-        'model_state_dict': training_dict['model'].state_dict(),
-        'optimiser_state_dict': training_dict['optimiser'].state_dict(),
-        'scheduler_state_dict': training_dict['lr_scheduler'].state_dict(),
-        'train_loss': training_dict['train_logger'],
-        'valid_loss': training_dict['validation_logger']
-    }, training_dict['save_path'])
-    print(f'{training_dict["dims"]}D - Epoch: {epoch}, Training Loss: {training_loss:.3f}, '
+    print(f'Epoch: {epoch}, Training Loss: {training_loss:.3f}, '
           f'Validation Loss: {validation_loss:.3f}, Time: {elapsed_time:.2f}')
-    print(f'LR Adjusted to {training_dict["lr_scheduler"].get_lr()[0]:.3f}')
+    lr = training_dict["lr_scheduler"].get_last_lr()[0]
+    if lr > 0:
+        print(f'{training_dict["dims"]}D - LR: {lr:.3f}, ', end='')
+
+    # If quit file exists, gracefully exit
+    if os.path.exists('terminate.txt'):
+        os.remove('terminate.txt')
+        # Save the model
+        torch.save({
+            'epoch': epoch,
+            'model_state_dict': training_dict['model'].state_dict(),
+            'optimiser_state_dict': training_dict['optimiser'].state_dict(),
+            'scheduler_state_dict': training_dict['lr_scheduler'].state_dict(),
+            'train_loss': training_dict['train_logger'],
+            'valid_loss': training_dict['validation_logger']
+        }, training_dict['model_path'])
+        print("\nUser requested termination of program. \n"
+              f"{training_dict['dims']}D model successfully saved after epoch {epoch}")
+        quit()
